Moved unit tests to new module for PostProcessingPlugin.
